Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    int radius;
-    scanf("%d", &radius);
-    int Area = 3.14 * radius * radius;
-    printf("Area: %d", Area);
+    float radius;
+    scanf("%f", &radius);
+    float Area = 3.14 * radius * radius;
+    printf("Area: %f", Area);
     return 0;  
 }
